requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
attrs==25.3.0
boto3==1.39.15
botocore==1.39.15
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
minio==7.2.16
numpy==2.3.2
oauthlib==3.3.1
outcome==1.3.0.post0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.23.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.13.1
schedule==1.2.2
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
